requirements.txt
discord.py[voice]==2.3.2
gspread==5.7.2
google-auth==2.16.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0